+++ src/main/java/io/blt/gregbot/ui/Ui.java
@@ -12,7 +12,7 @@ import com.formdev.flatlaf.FlatDarculaLaf;
 import io.blt.gregbot.ApplicationProperties;
 import io.blt.gregbot.ApplicationResources;
 import io.blt.gregbot.ui.dialogs.SplashScreen;
-import io.blt.gregbot.ui.forms.MainForm;
+import io.blt.gregbot.ui.frames.MainFrame;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -47,8 +47,8 @@ public class Ui {
             new SplashScreen("/splash/octogreg/2048.png", 1500)
                     .setVisible(true);
 
-            var mainForm = new MainForm();
-            mainForm.setVisible(true);
+            var mainFrame = new MainFrame();
+            mainFrame.setVisible(true);
         } catch (Exception e) {
             log.error("Unexpected exception", e);
         }
