+++ src/main/java/io/blt/gregbot/ui/frames/MainFrame.java
@@ -6,7 +6,7 @@
  * file, You can obtain one at: https://www.gnu.org/licenses/gpl-3.0.txt
  */
 
-package io.blt.gregbot.ui.forms;
+package io.blt.gregbot.ui.frames;
 
 import com.formdev.flatlaf.util.SystemInfo;
 import io.blt.gregbot.ApplicationProperties;
@@ -22,12 +22,12 @@ import java.awt.*;
 import java.awt.event.ActionListener;
 import javax.swing.*;
 
-public class MainForm extends JFrame {
+public class MainFrame extends JFrame {
 
     private JPanel contentPane;
     private LogbackPane logbackPane;
 
-    public MainForm() {
+    public MainFrame() {
         setJMenuBar(buildMenuBar());
         setContentPane(contentPane);
